refactor: ChatChannelReaderActor 메시지 분리
- 포트 역할을 수행하는 메시지를 domain/chat/port/out/ChatChannelReaderProtocol로 분리
