+++ pekko/src/test/java/com/tok/pekko/domain/chat/model/ChatChannelReaderActorTest.java
@@ -11,15 +11,15 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.timeout;
 import static org.mockito.Mockito.verify;
 
+import com.tok.pekko.domain.chat.port.in.ChatChannelReaderProtocol.ChatChannelReaderCommand;
+import com.tok.pekko.domain.chat.port.in.ChatChannelReaderProtocol.HistoryLoaded;
+import com.tok.pekko.domain.chat.port.in.ChatChannelReaderProtocol.RequestHistory;
+import com.tok.pekko.domain.chat.port.in.ChatChannelReaderProtocol.Shutdown;
+import com.tok.pekko.domain.chat.port.in.ChatChannelReaderProtocol.SyncNewCommand;
 import com.tok.pekko.domain.chat.port.out.ClientSessionProtocol;
 import com.tok.pekko.domain.chat.port.out.ClientSessionProtocol.ClientSessionCommand;
 import com.tok.pekko.domain.chat.port.out.ClientSessionProtocol.DeliverCommand;
 import com.tok.pekko.domain.chat.port.out.ClientSessionProtocol.DeliverHistory;
-import com.tok.pekko.domain.chat.model.ChatChannelReaderActor.ChatChannelReaderCommand;
-import com.tok.pekko.domain.chat.model.ChatChannelReaderActor.RequestHistory;
-import com.tok.pekko.domain.chat.model.ChatChannelReaderActor.HistoryLoaded;
-import com.tok.pekko.domain.chat.model.ChatChannelReaderActor.Shutdown;
-import com.tok.pekko.domain.chat.model.ChatChannelReaderActor.SyncNewCommand;
 import com.tok.pekko.domain.chat.port.out.MessageStoragePort;
 import java.time.Duration;
 import java.time.LocalDateTime;
